refactor(clippy): 工作空间 clippy 全清零 + 固化 workspace lint 策略
最终目标：`cargo clippy -p oc-core -p oc-server --all-targets -- -D warnings`
零 warning 通过，解锁 CI 接入 clippy --deny warnings。

步骤拆解：

1. 根 Cargo.toml 新增 [workspace.lints.clippy]（~40 条 allow，每条注释原因），
   三个 crate 一行 `[lints] workspace = true` 即对齐。allow 分三档：
   - 结构性 / refactor: too_many_arguments / type_complexity
   - 纯风格 idiom modernization（100+ 命中、与 correctness 无关）:
     map_or / redundant_closure / while_let_on_iterator / collapsible_if /
     needless_borrow / manual_{range_contains,clamp,div_ceil,is_multiple_of,strip} /
     explicit_counter_loop / needless_range_loop / if_same_then_else /
     let_unit_value / single_match_else / unnecessary_{to_owned,lazy_evaluations} /
     unused_async / useless_format / clone_on_copy / option_map_unit_fn /
     bind_instead_of_map / extra_unused_lifetimes / bool_comparison /
     needless_pass_by_ref_mut / redundant_field_names / single_char_pattern /
     only_used_in_recursion / match_single_binding / unnecessary_map_or /
     map_unwrap_or / map_clone / should_implement_trait / doc_lazy_continuation
   - 人工 audit 后 at-site-safe（crate-wide 允许）:
     module_inception / await_holding_lock / new_without_default /
     large_enum_variant

2. 手工修剩下 ~100 条真 warning:

   a) #[derive(Default)] 替换手写 impl Default — 16 个类型: ChannelHealth /
      DmPolicy / GroupPolicy / PlanModeState / Outcome / DeferredToolsConfig /
      EmbeddingProviderType / EmbeddingConfig / ImageGenProviderEntry /
      SkillStatus / SsrfPolicy / AutoReviewPromotion / ThinkingStyle /
      ProxyMode / ProxyConfig / ToolExecContext（非首变体的用 #[default]
      属性标注）

   b) items after test module 两处 — process_registry / system_prompt::build
      把测试模块外的 items 挪回 test mod 前方

   c) &mut Vec<T> -> &mut [T] 签名降级 — context_compact::{prune_old_context,
      compact_if_needed} / memory::embedding::utils::l2_normalize（函数体内
      无 push/pop/drain 操作，slice 签名足够）

   d) iter().any() -> contains() — file_extract / skills::author 4 处
      sort_by(a.cmp) -> sort_by_key — tools::ls
      u64 as u64 冗余 cast 删除 — telegram / wechat polling
      空循环迭代器 discard -> 直接 iter — subagent::spawn / image_generate::minimax
      手写 += -> += 运算符 — cron::db
      match { "a"|"b"|... => { return; }, _ => {} } -> if let Some("a"|"b"|...)
      — slack::socket
      find(|c| c=='&'||c==' '||...) -> find(['&',' ',...]) — logging::file_ops
      !A||!(B&&C) -> !(A&&B&&C) — skills::auto_review::triggers
      trim_start() 紧跟 split_whitespace() 多余，删 — skills::author

   e) 可见性 / dead_code 清理:
      - format_prompt_summary 去 #[deprecated]（3 个 caller 都是"单 budget
        转发 v2"场景，保留为 convenience wrapper 更清晰）
      - api_types 模块 pub(super)，ReasoningConfig pub(crate)（解决
        return type 比方法更私的警告）
      - AcpSessionStore 补 is_empty()
      - SqliteMemoryBackend::open 的 std::mem::transmute 加显式类型签名
      - weather_location_macos 的 extern_protocol! unsafe trait 走
        #![allow(clippy::missing_safety_doc)] crate-level（宏体里无处挂 doc）
      - channel::qqbot last_seq = 0 加 #[allow(unused_assignments)] 块 +
        注释"显式重置语义"
      - channel::{discord/feishu/slack/telegram} 的 bot_id / bot_name
        等 diagnostics 字段加 #[allow(dead_code)] + 注释
        "保留作为 future bot-authored event filtering"
      - build_static_system_prompt / tool_context / is_tool_denied 等
        "命名对偶的备用 API" 加 #[allow(dead_code)] + 注释

3. 顺带修了 telegram/sandbox 两个 clippy error（已在前一个 commit 解锁
   clippy 接入）。

测试: 455 passed（oc-core）+ 9 passed（oc-server）全绿。
parallel 工作线的 uncommitted 改动（agent/mod.rs / anthropic.rs + 3 个 orphan
文件）不在本 commit 范围，未 touch。

+++ crates/oc-core/src/channel/telegram/polling.rs
@@ -79,7 +79,7 @@ pub async fn run_polling_loop(
                         Err(e) => {
                             consecutive_errors += 1;
                             let backoff = std::cmp::min(
-                                (2u64.pow(consecutive_errors.min(5))) as u64,
+                                2u64.pow(consecutive_errors.min(5)),
                                 max_backoff_secs,
                             );
                             // Log first 3 errors as warn, then only every 10th to avoid spam
