+++ crates/oc-core/src/channel/wechat/polling.rs
@@ -48,7 +48,7 @@ pub(crate) async fn run_polling_loop(
                 break;
             }
             result = tokio::time::timeout(
-                std::time::Duration::from_millis(next_timeout_ms as u64 + 15_000),
+                std::time::Duration::from_millis(next_timeout_ms + 15_000),
                 api.get_updates(&sync_buf, Some(next_timeout_ms))
             ) => match result {
                 Ok(inner) => inner,
